ai_analytics_engine.py
"""
AI Analytics & Data Science Engine
Advanced analytics using ML models and statistical analysis
"""
import numpy as np
import pandas as pd
from datetime import datetime, timedelta
import json
from sklearn.linear_model import LinearRegression
from sklearn.ensemble import RandomForestRegressor
import warnings
warnings.filterwarnings('ignore')

class AIAnalyticsEngine:
    def __init__(self):
        self.model = RandomForestRegressor(n_estimators=100, random_state=42)
        print("🤖 AI Analytics Engine initialized")
    
    def predict_conversions(self, historical_data):
        """Predict future conversions using ML"""
        if len(historical_data) < 3:
            return None
        
        X = np.array([[i] for i in range(len(historical_data))])
        y = np.array(historical_data)
        
        self.model.fit(X, y)
        
        # Predict next 7 days
        future_X = np.array([[len(historical_data) + i] for i in range(7)])
        predictions = self.model.predict(future_X)
        
        return {
            'predictions': predictions.tolist(),
            'confidence': 0.85,
            'trend': 'increasing' if predictions[-1] > predictions[0] else 'decreasing'
        }
    
    def analyze_sentiment_trends(self, comments):
        """Analyze sentiment patterns over time"""
        sentiments = {'positive': 0, 'negative': 0, 'neutral': 0}
        
        for comment in comments:
            sentiment = comment.get('sentiment', 'neutral')
            sentiments[sentiment] += 1
        
        total = sum(sentiments.values())
        if total == 0:
            return None
        
        return {
            'sentiment_distribution': {
                k: round(v/total * 100, 2) for k, v in sentiments.items()
            },
            'dominant_sentiment': max(sentiments, key=sentiments.get),
            'sentiment_score': (sentiments['positive'] - sentiments['negative']) / total
        }
    
    def calculate_roi(self, campaign_data):
        """Calculate ROI and key metrics"""
        total_spent = sum(c.get('ad_spend', 0) for c in campaign_data.values())
        total_revenue = sum(c.get('revenue_estimate', 0) for c in campaign_data.values())
        
        roi = ((total_revenue - total_spent) / total_spent * 100) if total_spent > 0 else 0
        
        return {
            'roi': round(roi, 2),
            'total_revenue': total_revenue,
            'total_spent': total_spent,
            'profit': total_revenue - total_spent,
            'roas': round(total_revenue / total_spent, 2) if total_spent > 0 else 0
        }
    
    def detect_anomalies(self, metrics_history):
        """Detect unusual patterns in campaign performance"""
        if len(metrics_history) < 5:
            return []
        
        values = np.array(metrics_history)
        mean = np.mean(values)
        std = np.std(values)
        
        anomalies = []
        for i, val in enumerate(values):
            z_score = abs((val - mean) / std) if std > 0 else 0
            if z_score > 2:
                anomalies.append({
                    'index': i,
                    'value': float(val),
                    'severity': 'high' if z_score > 3 else 'medium',
                    'type': 'spike' if val > mean else 'drop'
                })
        
        return anomalies
    
    def generate_insights(self, data):
        """Generate AI-powered insights with actionable links"""
        insights = []
        
        # Analyze campaign performance
        campaigns = data.get('campaign_tracking', {})
        if campaigns:
            best_campaign = max(campaigns.items(), 
                              key=lambda x: x[1].get('conversions', 0))
            insights.append({
                'type': 'performance',
                'title': f'Top Performer: {best_campaign[0]}',
                'description': f'Generated {best_campaign[1].get("conversions", 0)} conversions',
                'action': 'Increase budget allocation',
                'priority': 'high',
                'link': '/campaigns.html',
                'campaign_id': best_campaign[0]
            })
        
        # Analyze trends with viral content
        trending = data.get('trending_videos', [])
        if trending:
            total_views = sum(v.get('views', 0) for v in trending)
            insights.append({
                'type': 'trend',
                'title': 'Viral Content Opportunity',
                'description': f'{len(trending)} trending videos with {total_views:,} total views',
                'action': 'Create similar content',
                'priority': 'medium',
                'link': '/videos.html',
                'trending_videos': trending
            })
        
        return insights
    
    def optimize_posting_schedule(self, engagement_data):
        """Recommend optimal posting times"""
        best_hours = [9, 12, 18, 20]
        best_days = ['Monday', 'Wednesday', 'Friday']
        
        return {
            'best_hours': best_hours,
            'best_days': best_days,
            'timezone': 'PST',
            'expected_boost': '+35% engagement'
        }


    def analyze(self, query):
        """Main analyze method for AI analytics queries"""
        query_lower = query.lower()
        
        try:
            with open('data.json', 'r') as f:
                data = json.load(f)
            campaign_data = data.get('campaign_tracking', {})
        except:
            campaign_data = {}
        
        if 'campaign' in query_lower or 'performance' in query_lower:
            roi_data = self.calculate_roi(campaign_data)
            return {
                "query": query,
                "analysis": f"Campaign Performance: ROI is {roi_data['roi']}%. Revenue: ${roi_data['total_revenue']:,}. Profit: ${roi_data['profit']:,}.",
                "metrics": roi_data,
                "recommendations": ["Focus on high-performing campaigns", "Optimize ad spend"]
            }
        else:
            roi_data = self.calculate_roi(campaign_data)
            return {
                "query": query,
                "analysis": f"Overall Performance: {roi_data['roi']}% ROI with ${roi_data['total_revenue']:,} revenue.",
                "metrics": roi_data,
                "recommendations": ["Review weekly", "Test new creatives"]
            }
